requirements.txt
google-generativeai
ollama
langchain
langchain-community
langchain-ollama
langchain-huggingface
python-dotenv
colorama
pypdf
python-docx
docx2txt
